+++ provisioning/tasks/bootstrap.sh
@@ -6,7 +6,6 @@ set -e
 # Make sure this script is executed by a superuser account.
 if [[ $(id -u) -ne 0 ]] ; then
     echo "This script must be executed by a superuser account (UID=0), aborting!"
-
     exit 1
 fi
 
@@ -16,6 +15,7 @@ export DEBIAN_FRONTEND=noninteractive
 
 apt-get update --allow-releaseinfo-change
 
-apt-get --yes \
-        install wget \
-                gpg
+apt-get --yes install \
+        curl \
+        wget \
+        gpg